+++ ICP/TrICP.h
@@ -30,8 +30,8 @@ private:
     float _smallTrimemdMSE;
 
     bool iterate(
-            Eigen::Matrix3d& R,
-            Eigen::Vector3d& t,
+            Eigen::Matrix3f& R,
+            Eigen::Vector3f& t,
             float& error,
             float eps = 0.001);
 };
